+++ tests/pallas/export_back_compat_pallas_test.py
@@ -48,9 +48,10 @@ class CompatTest(bctu.CompatTestBase):
       self.skipTest("Only works on GPUs with capability >= sm80")
     super().setUp()
 
-  @unittest.skip("TODO(necula): This test is checking backwards compatibility "
+  @unittest.skip("This test is checking backwards compatibility "
                  "of Triton IR, but Triton doesn't promise backwards "
-                 "compatibility for its IR.")
+                 "compatibility for its IR, and we have since removed "
+                 "the corresponding custom call from the guaranteed stable list.")
   def test_triton_add_one(self):
     def func(x):
       def add_one(x_ref, o_ref):